+++ packages/styles/src/index.ts
@@ -32,5 +32,6 @@ export { alert } from "./packages/alert";
 export { tabs } from "./packages/tabs";
 
 export type { BoxSprinkle } from "./packages/box";
+export type { IconButtonSprinkle } from "./packages/iconButton";
 export type { SidebarSprinkle } from "./packages/sidebar";
 export type { StackSprinkle } from "./packages/stack";