+++ packages/styles/src/packages/iconButton/index.ts
@@ -1,5 +1,17 @@
-import { style } from "./iconButton.style.css";
+import * as style from "./iconButton.style.css";
+import {
+  backgroundColorProperties,
+  borderColorProperties,
+  sprinkle,
+} from "./iconButton.sprinkle.css";
 
 export const iconButton = {
   style,
+  sprinkle,
+  properties: {
+    backgroundColor: backgroundColorProperties,
+    borderColor: borderColorProperties,
+  },
 };
+
+export type { IconButtonSprinkle } from "./iconButton.sprinkle.types";